repo/face-recognition-service/requirements.txt
flask==3.0.3
numpy==1.26.4
Pillow==10.3.0
